requirements.txt
python-doctr[torch]
rapidfuzz==2.15.1
